Add command line arg processing for the ccracker utiltiy.
Turns out the clap crate doesn't support non-unit enums. As a work
around, the dictionary data required by the dictionary attack is now
stored as part of the binary. This makes it so the user doesn't need to
specify a dictionary and the Attack::Dictionary variant doesn't need to
store the user's dictionary path.